updated frontend setup and added exit status

+++ roboshop/components/frontend.sh
@@ -2,8 +2,17 @@ echo frontend setup
 
 
 yum install nginx -y
+if [ $? -ne 0 ]; then
+  echo "nginx install failed"
+  exit 1
+fi
 
-curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zip"
+
+curl -s -L -o /tmp/frontend.zip "https://github.com/roboshop-devops-project/frontend/archive/main.zi"
+if [ $? -ne 0 ]; then
+  echo "download failed"
+  exit 1
+fi
 
 cd /usr/share/nginx/html
 rm -rf *
